workflows/requirements.txt
anthropic>=0.25.0
requests>=2.31.0
yfinance>=0.2.36
python-dotenv>=1.0.0
flask>=3.0.0
